Added some counter designs for investigating the DMUX 1:1 and ROACH 2 clock issue.
Both designs clock the fabric using adc0_clk which can be connected to either a
DMUX 1:1 and DMUX 1:2 board since the data lines are unused. There is a counter
connected to the LED's which counts up by one every ~40 seconds or so.

On ROACH 2 with a DMUX 1:2 board connected we get correct behaviour:

>> roach2.progdev('adc5g_r2_counter_2011_Nov_17_1718.bof')
>> floor(roach.est_brd_clk())
125.0

And not surprisingly the LEDs count up as expected and the F19/F20 pins on the
ZDOK show clock activity of the right frequency. However, when a DMUX 1:1 board
is attached, and the FPGA is programmed with the exact same bitcode, we see
the following behavior:

>> roach2.progdev('adc5g_r2_counter_2011_Nov_17_1718.bof')
>> floor(roach.est_brd_clk())
0.0

The LEDs do not count and the F19/F20 pins show no activity whatsoever.

Note: On the ROACH 1 both boards show the correct and expected behavior (using
the ROACH 1 bitcode of course).