Added the 12/16 bit to 4 bit quantizers.
These are currently implemented as lookup-tables and the scripts
to populate them are,

  quant12_hw_lut.m for the phase component
  quant16_hw_lut.m for the mag-squared component

These may change in the future to both be 16-bits or some other
bitwidth.

The reason for going with LUTs is for the flexibility of the
setting the quantization levels. This may be deemed unnecessary
in the future and may be replaced with simpler logic.