Added a White Gaussian Noise source.
This is based on a similar system designed by Dave MacMahon for the
SMA phased array processor. The block box consists of the following \
blocks

  noise_core.mdl
  noise_core.vhd
  noise_core/noise_co
  noise_core_config.m

Some changes to the WGN generator block were needed to get this
to work on the Virtex 6. These are described in the black box model
file.

Also added some place holders for the Walsh and complex gain control
logic, basically just added the registers needed.