Added the polar co-ordinate X-engine black-box core.
  xeng_core.mdl
  xeng_core.vhd
  xeng_core/xeng_core.ngc
  xeng_core_config.m

This X-engine cross multiplies and accumulates two inputs dual-
polarization and expects the F-data to be in polar form. It uses
75% less multipliers than its Cartesian counterpart at the expense
of more memory (due to the final polar-to-cartesian transformation
before accumulation).

It remains to be seen whether this will be the final X-engine. We
may find that we're memory limited and will have to move back to
Cartesian F-data, in which both this core and the complex gain core
will need to be modified.