Added single baseline X-engine and reduced clock rate to 143 MHz.
This single baseline correlator runs at half the rate of the final
design. This will allow debugging of most of the logic before the
addition of the packetization. Once the packetization logic is added
and debugged at this rate the full design will be retimed and floor-
planned at the faster 286 MHz fabric clock rate.