Added a mask to the X-engine block.
This masks out 16 of the 144 components for a total of 128. It is
currently set (it's not dynamically adjustable) to remove the cross-
pol products of the auto-correlations. This will need to change when
I figure out the correct order of the output.

Compiled using sma-wideband/mlib_devel@e05d91d.