Added sideband separation to the final accumulator.
The de-Walshing has yet to be implemented but this design now uses
two QDR accumulators to integrate the sidebands separately. The side-
band state is programmable and stored in the xengine_final_acc_sb_state
BRAM. The sync signal is assumed to indicate the start of a Walsh
period (i.e. it's the SOWF signal) and for each baseline and Walsh
step the 90-270 state is taken from sb_state and used to separate
sidebands.

To be consistent with the changes above the overall sync period has
been set to the SOWF period; however, since we're running at half-rate
this isn't really true. In fact, this breaks the integral number of
PFB windows per Walsh state requirement but will have to remain for now.

Additionally, since there is not enough block memory to store all
the visibilities they are instead sent out over two 10 GbE links on
every dump. The first 32 bins however are captured for debugging
purposes.