Changed how the X-engine is fed for the single-ROACH bench case.
The design now processes 1/8 of the channels as it normally would
(e.g. one of the parallel streams is processed) but it demultiplexes
them and sends 1/8 to each of the eight parallel X-engines. This has
the side-effect of changing which antenna is non-zero every channel
for a period of 8 channels in the final accumulator.

Keep in mind that this will change when the packetizers are added.
This is only temporary for the single-ROACH, auto-correlation bench
test.