Fix off-by-one bug in net rx fifo readout
The fix is to remove the delay block in network/eth/re_control.  This delay was
intended to match the latency of the counter, but it turns out that matching
the latency caused the initial relo output pulse to be one clock cycle shorter
than desired.  This initial one-clock-cycle-short relo pulse caused the data to
be shifted by one cycle forever after.  Removing this delay made things behave
as expected.