Capture network SOWF bit with first data word only
The SOWF bit from the packet header is high for all words of a sync packet.
Capturing it in the RX FIFOs with each 64-bit data word meant that it would be
toggling on the output of the RX FIFOs since they readout the data in 32-bit
"half words".  This can lead to problems of determining the "real" SOWF bit
downstream.  To fix this, we only capture the SOWF bit with the first 64 bit
data word of a sync packet.