Enable net RX FIFOs only after SOWF rising edge
If the RX FIFOs were enabled when the SOWF bit high (i.e. during the reception
of a sync packet), only a portion of the first packet would be written to the
FIFOs.  This would throw off the readout logic which expects only complete
packets to be written to the FIFOs.  The fix is to enable to write enables only
on the first rising edge of SOWF (after the "RX enable" bit is set in the
control register) rather than on any high cycle of SOWF.