Move EOF bit to low half of net RX words
The EOF (end of frame) bit is asserted with the last 64 bit word of the packet.
The network RX FIFOs store the SOWF and EOF bits alongside the data.  These two
bits of metadata need two extra bits of padding on the FIFO inputs since the
readout side of the FIFOs is only 32 bits wide.  The FIFO input used to be
constructed as:

SOWF, EOF, data[63:32], 2'b00, data[31:0]

The above arrangement associates both SOWF and EOF with the upper (i.e. first)
half of the 64-bit data word.  This is correct for SOWF, but EOF is really
associated with the lower (i.e. last) half of the 64-bit data word.  The new
arrangement is:

SOWF, 1'b0, data[63:32], 1'b0, EOF, data[31:0]