Extend SOWF so that all X engines receive it
The mcount generator also generates the SOWF signal for the data packets.  This
generated SOWF signal was high for 128 cycles which corresponded to the first
packet sent out each interface for a Walsh Frame.  The first packet sent out
each interface is always destined for the ROACH2 with FID=0, so only that
ROACH2 would get packets with SOWF=1 in the header.

To fix this, the generated SOWF is now held high for 8 times longer so that the
first 8 packets sent out an interface for a Walsh Frame have SOWF=1 in the
header.  This will allow all 8 ROACH2s to receive SOWF=1 in the header of the
appropriate packets.