Be smarter about SOWF receive buffering
The SOWF bit is asserted at t=0 for all FIDs when stored in the receive buffer.
Since SOWF becomes the X engine sync pulse, we must ensure that we output it
only once per Walsh frame (i.e. instead of once per FID per Walsh frame).
Additionally, we need to output it at the start of Walsh frame (i.e. with FID
0) even if we did not receive any FID 0 data.

This commit adds special buffering of the SOWF bit that is intended to
accomplish the above goals.