Change packet buffer readout control logic
The packet buffer readout control logic used to start reading from the packet
buffer when they had recevied half their capacity in packets.  This would be
fine when running with the full complement of ROACH2s, but when running with
half or a quarter of the ROACH2, this approach is flawed.

The new approach now starts reading from the packet buffers as soon as the
third of the four slots (i.e. the first slot of the second half) is written to.
This will allow start the readout at the apropriate time even when runninf with
fewer than the full complement of ROACH2s.