Add sma_corr_2014_Jan_16_0340.bof.gz
This version added the network/sync_count register that counts sync pulses
being sent to the X engines.

WARNING: THIS BUILD FAILED TO MEET TIMING!!!

This build failed timing, so is not usable for production, but I think the
timing errors will affect only the data values and NOT the sync pulse timing,
so it might be useful for testing whether the X engines are getting sync pulses
at the expected rate.