Add sma_corr_2014_Jan_18_1927.bof.gz
This BOF file has fixes to address the antenna mis-labeling bug (that is
suspected to be caused by extraneous SOWF/sync pulses being sent to the X
engines).  It also contains changes to address a weakness in the receive buffer
readout control logic when running with fewer than the full complement of
ROACH2s.