[sma_corr] Set the sample rate to the "psuedo" half-rate of 2496 MHz.
This is of course commensurate with our Walsh frequncy. This required
setting the clock rate of both ADCs to 1248 MHz and setting the XSG
core config clock rate to 156 MHz.

Additionally this meant changing the vectors_per_wstep setting from 11
to 6, meaning we're no longer using the even Walsh step trick. Hopefully
this will work!