[cgain_core] Fixed internal overflow issues causing strange amplitude fluctuations.
Also added four (4) fractional bits to the complex gain processing. This should allow
for full control with the full shifting of the FFT. Some internal bitwidths had to
change too, most notably the 'a' input of the DSB48E_1 has been set to the largest
allowable bitwidth, 25-bits. This also meant we had to swap the 'a' and 'b' inputs
which in turn meant the output had to be conjugate to be consistent with previous
versions.