[xeng_core] Resynthesized again with standard options.
This should not change anything! However there is seemingly a bug in the latest
bitcode that has added an unexpected cycle of latency somewhere in the X-engine.
If this still exists after a recompile then at least it's consistent and we start
to try and figure out where it's coming from.