Make B engine gain dual-pol
`beng_gain[ 7:0]` is the UFix_8_4 gain for the "X" polarization.
`beng_gain[15:8]` is the UFix_8_4 gain for the "Y" polarization.

This also rearranges the slicing in the summing blocks so that each SIMD vacc
handles a single polarization.  This will allow for future "per-antpol"
disabling (i.e. excluding) of antpols from the phased sums.