Add simd_test.mdl file
This is a model for testing the simd_vacc block.  At this point there are no
automated test script for it; only manual testing is supported.

This version includes the `disable` bits that can selectivly disable inputs.
Setting the LSb of `disable` to 1 will disable input 0.  In general,
`disable[i] == 1` will disable input `i`.