Add per-antpol disabling of B engine inputs
This adds the ability to exclude individual antpols from the phased sum.  The
lower 16 bits of the `beng_disable` register correspond to the 16 different
antpols.  A 1 bit will DISABLE the corresponding antpol so that it will not be
included in the phased sum.

`beng_disable[ 7:0]` are the disable bits for the "X" polarization.
`beng_disable[15:8]` are the disable bits for the "Y" polarization.

For `0 <= i <= 7`:

`beng_disable[  i]` will disable antenna `i`, polarization "X".
`beng_disable[8+i]` will disable antenna `i`, polarization "Y".