[fir_core] Re-introduced the non pass-through FIR netlist.
Somehow the pass-through netlist was brought in at commit @76368b1
and since then SWARM has been FFT-only. This brings in a re-compiled
FIR netlist that actually has a FIR. Now SWARM should once again be
PFB-based. Note: this netlist was also re-synthed with the standard
SWARM XST options using swarm_resynth_core.

Also, brought fir_core.mdl up-to-date and changed the FPGA clock
period parameter in the XSG block to 3.4965 ns (from the incorrect
value of 10 ns). I'm not sure if that last bit makes a difference
but it can't hurt.