[sma_corr] Decreased the sample rate to 10/11ths.
Hopefully this we can get this build to meeting timing, the efforts of
which may then be useful for the full rate case. In any event, presently the
corner turn will not work for 11/11ths so technically this is the maximum
clock rate we can really support; at least until we build in the loop-back
buffers.

This amounts to setting:

1. FPGA clock rate in XSG core config to 260.0 MHz
2. ADC clock rate to 2080 MHz
3. vectors_per_wstep to 10