[sma_corr] Since each XID now holds 2x more channels, network buffer reading needs to change.
Basically just reading the packet buffer out with a straight counter means that
the X-engines see data from the second set of channels from the previous antenna
instead of the first set of channels for the next antenna. So I decided to permute
the read address counter to re-arrange the data coming out in the order that the
X-engines expect.

Note 1: this permutation could have been done on the write address of the network
packet buffer (leaving the read address as-is) but since that address is used in
quite a few other places I chose to permute the read address instead.

Note 2: the permutation actually only moves a single bit around (bit 10) so any
logic that extracts the "slot number" (bits 11-12) from the read address will
still work as normal.