[sma_corr] Added logic which captues the real-time of four consecutive B-frames.
In the next commit this will be used in the B-engine header_gen block to
time tag the start of a B-engine frame as close to the samplers as possible.
Unfortunately since the latency of SWARM is as-of-yet unknown and may in fact be
multiple B-frames we need to capture the real-time of four consecutive B-frames
and determine which one is the valid by measuring the latency through other means.
Fortunately once this is determined it should be constant for all time.