[sma_corr] Changed the B-engine packet header to contain B-frame realtime.
The B-frame realtime value replaces the B-frame count which Dave
originally included in the header. The realtime value is still unique
to each complete B-frame and can be used in a similar fashion (however
it does not increase by 1 every time); however it can now also be used
to link the start of a B-frame with a certain absolute time.