[sma_corr] Added scan length to the visibility packet headers.
The internal scan length, xn_num, is 29-bits however this number
represents the number of 128-spectra to integrate for. Since this number
is generally a multiple of 64*11 I have stripped off the most-significant
24 bits and added it to the visibility packet header. The packet header now
looks like:

[MSB : 16-bits : packet # ][ 24-bits : scan # ][ 24-bits : scan length : LSB]

The listening code can now extract the scan length directly from the visib-
ility packets instead of reading it from each ROACH2. This is not only a more
direct (and hopefully) more accurate way to get the scan length but also should
speed up the listener code a bit and make it more robust.