[sma_corr] Redesign B-eng packetizer to DSB output
The packetizer now accepts beamformer output for two sidebands and
interleaves the packets that associate with each sideband. Data for
each packet are now clocked continually for the full duration of the
packet (whereas data was clocked at 2/3 duty cycle before).

The B-engine ethernet block has also been updated to support two
destination IP addresses, and takes an additional input that selects
which IP address to use for the current packet. The updated packetizer
alternates between the IP addresses so that packets for one sideband go
to one destination, and packets for the second sideband to another.

For now data from "two sidebands" are generated by simply duplicating
the data for one sideband in order to test the DSB packetizer. However,
the ordering of the two polarizations in the duplicate stream is flipped
to allow recognition of the second sideband stream.