[sma_corr] Store 2nd sideband phase in registers
The shared BRAM used to store phasing coefficients for the second
sideband has been replaced with four 32-bit software registers, each
storing the coefficients for two FIDs. Even-numbered FIDs (0..6) are
in the lower, and odd-numbered FIDs (1..7) are in the higher 16 bits
of each register. Registers are named ..fid10, ..fid32, ..fid54,
..fid76 to indicate which two FIDs are stored in each of them.