[sma_corr] Combine 2nd sideband demod and phase
The phasing coefficients and demodulation patterns for the second
sideband beamformer are now stored in a single shared BRAM. The data
are 32-bit wide and occupy 512 addresses in range 0x000-0x1FF. Each
32-bit word stores the demodulation phase (0/1 for 0/180 deg phase)
and a 16-bit beamformer phase (fractional cycle in range [0,1)). The
demodulation phase occupies the MSb and the beamformer phase the least
significant portion of the word:

   b31 | b30 ... b16 | b15       ... b0
 ------+-------------+-----------------
 demod |  unassigned | beamformer phase

A group-of-8 words comprise the phases for 8 FIDs in a single Walsh
step, and 64 groups-of-8 comprise the steps for an entire Walsh frame,
so the word at 0x000 stores the phase for FID0 in step0:

0x000 => FID0, step0
0x001 => FID1, step0
 :        :     :
0x007 => FID7, step0
0x008 => FID0, step1
 :        :     :
0x1FF => FID7, step63
0x200 => unassigned
 :    =>  :
0x3FF => unassigned.

Note that the shared BRAM has 1024x32-bit of memory, due to toolflow
constraints. The upper half of the address space is unused.