[sma_corr] Reduce 2nd sideband phase complexity
Phasing coefficients for the second sideband beamformer are now
provided as Fix_7_5 real and imaginary components instead of a single
UFix_16_16 phase. This reduces the shared BRAM data to 16-bit values,
for which the format is as follows:

   b15 | b14 | b13 ... b7 | b6 ... b0
 ------+-----+------------+----------
 demod | u/a |       real |      imag

The use of Fix_7_5 for the coefficient components allows precise
representation of +1.0.

Note that the shared BRAM block now has an address space that extends
to 0x7FF, of which the upper three quarters are unused.