[sma_corr] Revert to 2/3 duty cycle beng packetizer
The double sideband packetizer now implements two of the original
2/3 duty cycle packetizers in parallel. This reversion recovers all
of the detailed operation of the original packetizer in an attempt
to fix a bug in the current implementation which causes data to slip
in frequency over time (i.e. the frequency channels contained in a
packet do not always match the header signature).