Rebuild cores for -2 speed grade
The cores were rebuilt after changing the model files to use the -2
speed grade FPGA. This was followed by running swarm_resynth_core on
each of the cores. This seems to have improved the timing score of
the sma_corr build.

Note that some diagnostic elements (at the Simulink simulation level)
had to be removed, apparently to be compatible with a newer version
of Matlab than used to create the cores initially.