[sma_corr] Unfix commit 9603b67
Commit 9603b67 causes slipping in the B-engine data so that the
frequency bins move with respect to the channel parameter embedded
in the packet headers. This is due to the fact that the summation
block for the 1st sideband assumes that valid data at its output
only starts appearing when the acc_init signal goes high *after*
the first acc_init falling edge that occurs after the sync signal
goes high. That assertion is true at startup, but subsequent sync
pulses that arrive at the start of each Walsh frame cause the
summation block not to produce a valid high signal for the first
batch of summed data into each Walsh frame.

To account for this effect, the valid_gate block in the packetizer
ORed the valid signal with a 3000-clock delayed version of itself
to ensure that a single batch of valid data unaccompanied by a
valid high signal would still be passed through if the valid signal
went high 3000 clocks earlier. The mechanism works, but the delay
needs to be a multiple of 1024 to avoid the data misalignment
described in commit message 9603b67.