[sma_corr/walsh] Add 2nd SB demod-phase field to ctrl / status
Bits 13..10 of the control register now contains a field that sets
the phase of beng/sb1demodphase/count11 at which the next counter in
the hierarchy increments. The field is 4-bit wide and should be less
than 11.

Bits 15..10 of the status register now contains the Walsh-step of
the demodulator in the second sideband beamformer.